add react-router and pages
