+++ package.json
@@ -51,6 +51,7 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.52.1",
     "react-redux": "^9.1.2",
+    "react-router-dom": "^6.26.0",
     "redux-persist": "^6.0.0",
     "uuid": "^10.0.0"
   }
