+++ src/main.tsx
@@ -5,8 +5,6 @@ import { Provider } from 'react-redux';
 import { store, persister } from './services/store';
 import { PersistGate } from 'redux-persist/integration/react';
 
-// Application to Render
-
 const app = (
   <Provider store={store}>
     <PersistGate loading={null} persistor={persister}>
@@ -15,5 +13,4 @@ const app = (
   </Provider>
 );
 
-// Render application in DOM
 createRoot(document.getElementById('app')).render(app);
