+++ src/pages/HomePage.module.scss
@@ -1,4 +1,4 @@
-@import './global.scss';
+@import '../components/global.scss';
 
 .mainContainer {
   display: flex;
