+++ src/pages/Layout.module.scss
@@ -1,0 +1,6 @@
+@import '../components/global.scss';
+
+.mainContainer {
+  display: flex;
+  margin-top: 80px;
+}
