+++ src/pages/ViewNote.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ViewNote() {
+  return <div></div>;
+}
