+++ src/presentations/FormUi/DatePicker/DatePicker.module.scss
@@ -12,6 +12,7 @@
     border-color: var(--card-title-text) !important;
     box-shadow: 0 0 0 2px rgba(255, 201, 201, 0.3) !important;
   }
+
   &:focus-within {
     border-color: var(--card-title-text) !important;
     box-shadow: 0 0 0 2px rgba(255, 201, 201, 0.3) !important;
