--- src/presentations/Layout/Layout.tsx
@@ -1,13 +1,0 @@
-// import React, { FC } from 'react';
-// import { Header } from '../Header/Header';
-// import { Footer } from '../Footer/Footer';
-
-// export const Layout = ({ children }) => {
-//   return (
-//     <div id='container'>
-//       <Header />
-//       <div className='main-container'>{children}</div>
-//       <Footer />
-//     </div>
-//   );
-// };
