+++ src/presentations/Modal/Modal.module.scss
@@ -26,5 +26,5 @@
 }
 
 .modalTitle {
-  color: var(--card-title-text);
+  color: var(--text-color);
 }
